Refactor plugin operations with pluginStorage
pluginStorage implements a common interface to work with lists of
plugins and allows to check for updates, not installed plugins and
perform updates and installations. Steps with checks, updates and bare
installations are clearly separated now. This gives us clearer code in
shell commands.

To achieve that, the Plugin interface now have an additional IsInstalled
method, that only checks if a plugin directory exists.