Print update and installation notes on plugins load and refactor update
This required the ability to check for downloaded updates without
accessing the Internet. This ability is used on plugins load to almost
instantly check for already downloaded updates.

Additionally, this commit contains a refactoring of the update command
which makes the overall architecture more consistent. This refactoring
also disallows a user to pass a plugin that is not listed within the
configuration.